CARL ResourceManager: Rename `api` module to `types`.
