+++ opendut-carl/src/cli.rs
@@ -6,7 +6,7 @@ use opendut_model::cluster::{ClusterDescriptor, ClusterDeployment, ClusterId};
 use opendut_model::peer::{PeerDescriptor, PeerId};
 use opendut_model::peer::configuration::PeerConfiguration;
 use crate::CreateResult;
-use crate::resource::api::Resource;
+use crate::resource::types::Resource;
 
 const BANNER: &str = r"
                          _____     _______
