+++ opendut-carl/src/manager/cluster_manager/create_cluster_descriptor.rs
@@ -1,6 +1,6 @@
 use opendut_model::cluster::{ClusterDescriptor, ClusterId, ClusterName};
 use tracing::{debug, info};
-use crate::resource::api::resources::Resources;
+use crate::resource::types::resources::Resources;
 use crate::resource::persistence::error::PersistenceError;
 use crate::resource::storage::ResourcesStorageApi;
