+++ opendut-carl/src/manager/cluster_manager/delete_cluster_descriptor.rs
@@ -2,7 +2,7 @@ use opendut_model::cluster::ClusterDisplay;
 use opendut_model::ShortName;
 use opendut_model::cluster::{ClusterDescriptor, ClusterDeployment, ClusterId, ClusterName};
 use opendut_model::cluster::state::ClusterState;
-use crate::resource::api::resources::Resources;
+use crate::resource::types::resources::Resources;
 use crate::resource::persistence::error::PersistenceError;
 use crate::resource::storage::ResourcesStorageApi;
