+++ opendut-carl/src/manager/peer_manager/assign_cluster/mod.rs
@@ -9,7 +9,7 @@ use opendut_model::peer::{PeerDescriptor, PeerId};
 use opendut_model::util::net::{NetworkInterfaceDescriptor, NetworkInterfaceName, NetworkInterfaceNameError};
 use tracing::debug;
 use opendut_carl_api::carl::broker::{ApplyPeerConfiguration, DownstreamMessagePayload};
-use crate::resource::api::resources::Resources;
+use crate::resource::types::resources::Resources;
 
 
 pub(in crate::manager) use assignment::{ClusterAssignment, PeerClusterAssignment};
