+++ opendut-carl/src/manager/peer_manager/list_peer_member_states.rs
@@ -1,5 +1,5 @@
 use crate::manager::cluster_manager;
-use crate::resource::api::resources::Resources;
+use crate::resource::types::resources::Resources;
 use crate::resource::persistence::error::PersistenceError;
 use crate::resource::storage::ResourcesStorageApi;
 use opendut_model::cluster::ClusterId;
