+++ opendut-carl/src/manager/peer_manager/store_peer_descriptor.rs
@@ -1,4 +1,4 @@
-use crate::resource::api::resources::Resources;
+use crate::resource::types::resources::Resources;
 use crate::resource::persistence::error::PersistenceError;
 use crate::resource::storage::ResourcesStorageApi;
 use crate::settings::vpn::Vpn;
