+++ opendut-carl/src/manager/test_manager/delete_viper_source_descriptor.rs
@@ -1,7 +1,7 @@
 use opendut_model::{format::format_id_with_optional_name, viper::{ViperSourceDescriptor, ViperSourceId, ViperTestRunDescriptor, ViperTestId}};
 use tracing::debug;
 use opendut_model::viper::ViperTestSuiteIdentifier;
-use crate::resource::{api::resources::Resources, persistence::error::PersistenceError, storage::ResourcesStorageApi};
+use crate::resource::{types::resources::Resources, persistence::error::PersistenceError, storage::ResourcesStorageApi};
 
 
 impl Resources<'_> {
