+++ opendut-carl/src/manager/test_manager/get_viper_test_suite_parameters.rs
@@ -5,7 +5,7 @@ use opendut_viper_rt::events::emitter;
 use opendut_viper_rt::source::Source;
 use opendut_viper_rt::{ViperOptions, ViperRuntime};
 use opendut_viper_rt::source::loaders::HttpSourceLoader;
-use crate::resource::{api::resources::Resources, persistence::error::PersistenceError, storage::ResourcesStorageApi};
+use crate::resource::{types::resources::Resources, persistence::error::PersistenceError, storage::ResourcesStorageApi};
 
 
 impl Resources<'_> {
