--- opendut-carl/src/resource/api/mod.rs
@@ -1,6 +1,0 @@
-pub mod id;
-pub mod resources;
-pub mod subscription;
-mod types;
-
-pub use types::Resource;
